# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Portfolio OS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      description: List all public projects, optionally filtered
      parameters:
        - name: tag
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: classification
          in: query
          schema:
            type: string
        - name: isPublic
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create project
      description: Create a new project (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project
      description: Get a single project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      description: Update an existing project (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      description: Delete a project (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}/sections:
    get:
      operationId: listProjectSections
      tags: [sections]
      summary: List project sections
      description: Get all sections for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of sections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSection"
    post:
      operationId: createProjectSection
      tags: [sections]
      summary: Create project section
      description: Add a section to a project (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSectionBody"
      responses:
        "201":
          description: Created section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSection"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}/sections/{sectionId}:
    patch:
      operationId: updateProjectSection
      tags: [sections]
      summary: Update project section
      description: Update a project section (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: sectionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSectionBody"
      responses:
        "200":
          description: Updated section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSection"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProjectSection
      tags: [sections]
      summary: Delete project section
      description: Delete a project section (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: sectionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /vault/notes:
    get:
      operationId: listVaultNotes
      tags: [vault]
      summary: List vault notes
      description: Get all vault notes (authenticated only)
      responses:
        "200":
          description: List of vault notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VaultNote"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /vault/notes/{projectId}:
    get:
      operationId: getVaultNote
      tags: [vault]
      summary: Get vault note for project
      description: Get vault note for a specific project (authenticated only)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vault note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultNote"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertVaultNote
      tags: [vault]
      summary: Upsert vault note
      description: Create or update a vault note for a project (admin only)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertVaultNoteBody"
      responses:
        "200":
          description: Upserted vault note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultNote"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        category:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - createdAt
...
    Project:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        summary:
          type: string
        classification:
          type: string
        status:
          type: string
        year:
          type: ["integer", "null"]
        isPublic:
          type: boolean
        isFeatured:
          type: boolean
        techStack:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        coverImageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - slug
        - summary
        - classification
        - status
        - isPublic
        - isFeatured
        - techStack
        - tags
        - createdAt
        - updatedAt
    ProjectSection:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        type:
          type: string
        title:
          type: string
        content:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - type
        - title
        - content
        - sortOrder
        - createdAt
    ProjectDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        summary:
          type: string
        classification:
          type: string
        status:
          type: string
        year:
          type: ["integer", "null"]
        isPublic:
          type: boolean
        isFeatured:
          type: boolean
        techStack:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        sections:
          type: array
          items:
            $ref: "#/components/schemas/ProjectSection"
        media:
          type: array
          items:
            $ref: "#/components/schemas/MediaAsset"
        relatedProjects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        coverImageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - slug
        - summary
        - classification
        - status
        - isPublic
        - isFeatured
        - techStack
        - tags
        - sections
        - media
        - relatedProjects
        - createdAt
        - updatedAt
    CreateProjectBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        summary:
          type: string
        classification:
          type: string
        status:
          type: string
        year:
          type: ["integer", "null"]
        isPublic:
          type: boolean
        isFeatured:
          type: boolean
        techStack:
          type: array
          items:
            type: string
        tagIds:
          type: array
          items:
            type: integer
        coverImageUrl:
          type: ["string", "null"]
      required:
        - title
        - slug
        - summary
        - classification
        - status
    UpdateProjectBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        summary:
          type: string
        classification:
          type: string
        status:
          type: string
        year:
          type: ["integer", "null"]
        isPublic:
          type: boolean
        isFeatured:
          type: boolean
        techStack:
          type: array
          items:
            type: string
        tagIds:
          type: array
          items:
            type: integer
        coverImageUrl:
          type: ["string", "null"]
    CreateSectionBody:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        content:
          type: string
        sortOrder:
          type: integer
      required:
        - type
        - title
        - content
        - sortOrder
    UpdateSectionBody:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        content:
          type: string
        sortOrder:
          type: integer
    MediaAsset:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        type:
          type: string
        url:
          type: string
        caption:
          type: ["string", "null"]
        altText:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - type
        - url
        - sortOrder
        - createdAt
...
    VaultNote:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - content
        - createdAt
        - updatedAt
...
    UpsertVaultNoteBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...