# lib/api-zod/src/generated/types/projectDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio OS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MediaAsset } from "./mediaAsset";
import type { Project } from "./project";
import type { ProjectSection } from "./projectSection";
import type { Tag } from "./tag";

export interface ProjectDetail {
  id: number;
  title: string;
  slug: string;
  summary: string;
  classification: string;
  status: string;
  /** @nullable */
  year?: number | null;
  isPublic: boolean;
  isFeatured: boolean;
  techStack: string[];
  tags: Tag[];
  sections: ProjectSection[];
  media: MediaAsset[];
  relatedProjects: Project[];
  /** @nullable */
  coverImageUrl?: string | null;
  createdAt: Date;
  updatedAt: Date;
}