# lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio OS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProjectBody {
  title?: string;
  slug?: string;
  summary?: string;
  classification?: string;
  status?: string;
  /** @nullable */
  year?: number | null;
  isPublic?: boolean;
  isFeatured?: boolean;
  techStack?: string[];
  tagIds?: number[];
  /** @nullable */
  coverImageUrl?: string | null;
}