# lib/api-zod/src/generated/types/updateSectionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio OS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSectionBody {
  type?: string;
  title?: string;
  content?: string;
  sortOrder?: number;
}