# lib/api-zod/src/generated/types/upsertVaultNoteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio OS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpsertVaultNoteBody {
  content: string;
}