# lib/api-zod/src/generated/types/vaultNote.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio OS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VaultNote {
  id: number;
  projectId: number;
  content: string;
  createdAt: Date;
  updatedAt: Date;
}